+++ packages/frontend/src/app/(sticky-header)/search/load-more-results.tsx
